security

+++ src/main/java/epam/news/dao/NewsDAO.java
@@ -4,5 +4,5 @@ package epam.news.dao;
 import epam.news.model.entity.News;
 
 public interface NewsDAO extends BasicDAO<News> {
-    News findById(int newsId);
+    News findById(Long newsId);
 }
